+++ CHANGELOG.md
@@ -9,62 +9,90 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as straightforward as possible.
 
+### Changed 
+
+- Modified CHANGELOG.md, nodemon.json & tsconfig.json
+
 ## [0.1.6-beta] - 2022-06-20
 
+### Changes
+
+- [87d536b](https://github.com/KarthikUdyawar/url-shortener/commit/87d536b) - Patch change
+
 ### Removed
+
 - [68c3ac2](https://github.com/KarthikUdyawar/url-shortener/commit/68c3ac2) - Deleted IErrorMessage interface
+
 ### Fixed
+
 - [a641cbd](https://github.com/KarthikUdyawar/url-shortener/commit/a641cbd) - Updated RedirectShortUrl response system
 
 ## [0.1.5-beta] - 2022-06-20
+
 ### Added
+
 - [7828a80](https://github.com/KarthikUdyawar/url-shortener/commit/7828a80) - Create IResponse.ts
 - [876c368](https://github.com/KarthikUdyawar/url-shortener/commit/876c368) - Create IMessage.ts
+
 ### Changed
+
 - [1e3a219](https://github.com/KarthikUdyawar/url-shortener/commit/1e3a219) - Change test.controller -> Test.controller
+
 ### Deprecated
+
 - [c236c2b](https://github.com/KarthikUdyawar/url-shortener/commit/c236c2b) - IErrorMessage interface will be removed
+
 ### Removed
+
 - [a8e714a](https://github.com/KarthikUdyawar/url-shortener/commit/a8e714a) - Remove useless import and variable
 - [7c1fbba](https://github.com/KarthikUdyawar/url-shortener/commit/7c1fbba) - Delete ICreateReqBody interface
+
 ### Fixed
+
 - [836904e](https://github.com/KarthikUdyawar/url-shortener/commit/836904e) - Using IResponse & IMessage interfaces in response
 
 ## [0.1.4-beta] - 2022-06-20
 
 ### Fixed
+
 - [68dd928](https://github.com/KarthikUdyawar/url-shortener/commit/68dd928) - Use IReqBody instead of ICreateReqBody
 
 ### Removed
+
 - [7c1fbba](https://github.com/KarthikUdyawar/url-shortener/commit/7c1fbba) - Delete ICreateReqBody interface
 
 ## [0.1.3-beta] - 2022-06-20
 
 ### Fixed
+
 - [a051f22](https://github.com/KarthikUdyawar/url-shortener/commit/a051f22) - Fixed Create short URL API
 
 ## [0.1.2-beta] - 2022-06-20
 
 ### Fixed
+
 - [1735da9](https://github.com/KarthikUdyawar/url-shortener/commit/1735da9) - Fixed RedirectShortUrl API
 
 ---
+
 ## [0.1.1-alpha.2] - 2022-06-19
 
 ### Added
+
 - [62725e4](https://github.com/KarthikUdyawar/url-shortener/commit/62725e4) - Added Redirect short URL API
 - [89ea302](https://github.com/KarthikUdyawar/url-shortener/commit/89ea302) - Added Create short utl utl API
 - [174b9bf](https://github.com/KarthikUdyawar/url-shortener/commit/174b9bf) - Added shortURL model
 - [b45bee5](https://github.com/KarthikUdyawar/url-shortener/commit/b45bee5) - Added router and controller 🎉
- 
+
 ### Changed
+
 - [9723a7e](https://github.com/KarthikUdyawar/url-shortener/commit/9723a7e) - Downgrade major versions
 
 ## [0.1.0-alpha.1] - 2022-06-19
 
 ### Added
 
-- [f79f46a](https://github.com/KarthikUdyawar/url-shortener/commit/f79f46a) - Merge pull request #1 from KarthikUdyawar/add-license-1 
+- [f79f46a](https://github.com/KarthikUdyawar/url-shortener/commit/f79f46a) - Merge pull request #1 from KarthikUdyawar/add-license-1
 - [a8d8d09](https://github.com/KarthikUdyawar/url-shortener/commit/a8d8d09) - (origin/add-license-1) Create LICENSE
 - [e307afa](https://github.com/KarthikUdyawar/url-shortener/commit/e307afa) - Start server and test API