+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrading notes for brands. It's a team effort to make them as straightforward as possible.
 
-### Changed
-- Miner changes for some files
+### Added
+- Added cors and morgan for middleware
 
 ## [0.1.6-beta] - 2022-06-20